Refactor common parts
